requirements.txt
locust>=2.20.0
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.31.0
jsonschema>=4.20.0
